Unnest `VersionID` and move to separate file
